+++ ASFWDriver/Isoch/Transmit/IsochTxDescriptorSlab.hpp
@@ -52,13 +52,15 @@ public:
 
     [[nodiscard]] uint8_t* PayloadPtr(uint32_t packetIndex) noexcept {
         return bufRegion_.virtualBase
-            ? (reinterpret_cast<uint8_t*>(bufRegion_.virtualBase) + (packetIndex * Layout::kMaxPacketSize))
+            ? (reinterpret_cast<uint8_t*>(bufRegion_.virtualBase) +
+               (static_cast<size_t>(packetIndex) * Layout::kMaxPacketSize))
             : nullptr;
     }
 
     [[nodiscard]] const uint8_t* PayloadPtr(uint32_t packetIndex) const noexcept {
         return bufRegion_.virtualBase
-            ? (reinterpret_cast<const uint8_t*>(bufRegion_.virtualBase) + (packetIndex * Layout::kMaxPacketSize))
+            ? (reinterpret_cast<const uint8_t*>(bufRegion_.virtualBase) +
+               (static_cast<size_t>(packetIndex) * Layout::kMaxPacketSize))
             : nullptr;
     }
